LC-4sum [Runtime: 16 ms; Memory: 17.5 MB]

+++ sorting/4sum.cpp
@@ -18,7 +18,7 @@ public:
                 int right = nums.size() - 1;
 
                 while(left < right){
-                    sum = nums[left] + nums[right] + nums[i] + nums[j];
+                    sum = (long long)nums[left] + nums[right] + nums[i] + nums[j];
 
                     if(sum < target) left++;
                     else if(sum > target) right--;
